User can visit IDP after logging in with SP
**Why**: Session vars made it so that a user was always redirected to
the service provider after logging in, which is annoying.
